Changed coordinate types from s32 to s16 in normal sprite draw, this fixes sprites in the Sonic Jam demos. I guess other commands should use s16 as well. Please test and report problems.
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2003 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/vidsoft.c
@@ -1796,7 +1796,7 @@ void VIDSoftVdp1DistortedSpriteDraw() {
 void VIDSoftVdp1NormalSpriteDraw(void)
 {
    vdp1cmd_struct cmd;
-   s32 x0, y0, x1, y1, W, H;
+   s16 x0, y0, x1, y1, W, H;
    u8 flip, SPD, endCode;
    u16 colorbank;
    int type;
